+++ src/postscribeRender.js
@@ -1,0 +1,8 @@
+import postscribe from 'postscribe';
+
+export function writeAdHtml(markup) {
+    postscribe(document.body, markup, {
+      error: console.error
+    });
+  }
+
