+++ test/spec/nativeRenderManager_spec.js
@@ -66,7 +66,7 @@ describe('nativeRenderManager', function () {
         },
       }];
       let nativeTracker = new newNativeRenderManager(mockWin);
-      nativeTracker.renderNativeAd(tagData);
+      nativeTracker.renderNativeAd(mockWin.document, tagData);
 
       expect(mockWin.parent.postMessage.callCount).to.equal(1);
       let postMessageTargetDomain = mockWin.parent.postMessage.args[0][1];
@@ -93,7 +93,7 @@ describe('nativeRenderManager', function () {
       }];
 
       let nativeTracker = new newNativeRenderManager(mockWin);
-      nativeTracker.renderNativeAd(tagData);
+      nativeTracker.renderNativeAd(mockWin.document, tagData);
 
       expect(mockWin.parent.postMessage.callCount).to.equal(2);
